chore: separate buildSearchParams from buildQuery
