+++ src/mongoose/buildQuery.ts
@@ -4,19 +4,21 @@
 import deepmerge from 'deepmerge';
 import { FilterQuery } from 'mongoose';
 import { combineMerge } from '../utilities/combineMerge';
-import { operatorMap } from './operatorMap';
-import { sanitizeQueryValue } from './sanitizeQueryValue';
 import { PayloadRequest, Where } from '../types';
-import { Field, FieldAffectingData, fieldAffectsData, TabAsField, UIField } from '../fields/config/types';
+import { Field, FieldAffectingData, TabAsField, UIField } from '../fields/config/types';
 import { CollectionPermission, FieldPermissions, GlobalPermission } from '../auth';
 import QueryError from '../errors/QueryError';
-import { getLocalizedPaths } from './getLocalizedPaths';
+import { buildSearchParam } from './buildSearchParams';
 
-const validOperators = ['like', 'contains', 'in', 'all', 'not_in', 'greater_than_equal', 'greater_than', 'less_than_equal', 'less_than', 'not_equals', 'equals', 'exists', 'near'];
+export const validOperators = ['like', 'contains', 'in', 'all', 'not_in', 'greater_than_equal', 'greater_than', 'less_than_equal', 'less_than', 'not_equals', 'equals', 'exists', 'near'];
 
-const subQueryOptions = {
-  limit: 50,
-  lean: true,
+export type EntityPolicies = {
+  collections?: {
+    [collectionSlug: string]: CollectionPermission;
+  };
+  globals?: {
+    [globalSlug: string]: GlobalPermission;
+  };
 };
 
 export type PathToQuery = {
@@ -28,11 +30,6 @@ export type PathToQuery = {
   fieldPolicies?: {
     [field: string]: FieldPermissions
   }
-}
-
-type SearchParam = {
-  path?: string,
-  value: unknown,
 }
 
 type ParamParserArgs = {
@@ -150,7 +147,12 @@ export class ParamParser {
           if (typeof pathOperators === 'object') {
             for (const operator of Object.keys(pathOperators)) {
               if (validOperators.includes(operator)) {
-                const searchParam = await this.buildSearchParam({
+                const searchParam = await buildSearchParam({
+                  collectionSlug: this.collectionSlug,
+                  errors: this.errors,
+                  globalSlug: this.globalSlug,
+                  policies: this.policies,
+                  req: this.req,
                   fields: this.fields,
                   incomingPath: relationOrPath,
                   val: pathOperators[operator],
@@ -190,158 +192,6 @@ export class ParamParser {
       }
     }
     return completedConditions;
-  }
-
-  // Convert the Payload key / value / operator into a MongoDB query
-  async buildSearchParam({
-    fields,
-    incomingPath,
-    val,
-    operator,
-    overrideAccess,
-  }: {
-    fields: Field[],
-    incomingPath: string,
-    val: unknown,
-    operator: string
-    overrideAccess: boolean
-  }): Promise<SearchParam> {
-    // Replace GraphQL nested field double underscore formatting
-    let sanitizedPath = incomingPath.replace(/__/gi, '.');
-    if (sanitizedPath === 'id') sanitizedPath = '_id';
-
-    let paths: PathToQuery[] = [];
-
-    let hasCustomID = false;
-
-    if (sanitizedPath === '_id') {
-      const customIDfield = this.req.payload.collections[this.collectionSlug]?.config.fields.find((field) => fieldAffectsData(field) && field.name === 'id');
-
-      let idFieldType: 'text' | 'number' = 'text';
-
-      if (customIDfield) {
-        if (customIDfield?.type === 'text' || customIDfield?.type === 'number') {
-          idFieldType = customIDfield.type;
-        }
-
-        hasCustomID = true;
-      }
-
-      paths.push({
-        path: '_id',
-        field: {
-          name: 'id',
-          type: idFieldType,
-        },
-        complete: true,
-        collectionSlug: this.collectionSlug,
-      });
-    } else {
-      paths = await getLocalizedPaths({
-        errors: this.errors,
-        policies: this.policies,
-        req: this.req,
-        collectionSlug: this.collectionSlug,
-        globalSlug: this.globalSlug,
-        fields,
-        incomingPath: sanitizedPath,
-        overrideAccess,
-      });
-    }
-
-    const [{ path, field }] = paths;
-
-    if (path) {
-      const formattedValue = sanitizeQueryValue({
-        field,
-        path,
-        operator,
-        val,
-        hasCustomID,
-      });
-
-      // If there are multiple collections to search through,
-      // Recursively build up a list of query constraints
-      if (paths.length > 1) {
-        // Remove top collection and reverse array
-        // to work backwards from top
-        const pathsToQuery = paths.slice(1).reverse();
-
-        const initialRelationshipQuery = {
-          value: {},
-        } as SearchParam;
-
-        const relationshipQuery = await pathsToQuery.reduce(async (priorQuery, { path: subPath, collectionSlug }, i) => {
-          const priorQueryResult = await priorQuery;
-
-          const SubModel = this.req.payload.collections[collectionSlug].Model;
-
-          // On the "deepest" collection,
-          // Search on the value passed through the query
-          if (i === 0) {
-            const subQuery = await SubModel.buildQuery({
-              where: {
-                [subPath]: {
-                  [operator]: val,
-                },
-              },
-              req: this.req,
-              overrideAccess,
-            });
-
-            const result = await SubModel.find(subQuery, subQueryOptions);
-
-            const $in = result.map((doc) => doc._id.toString());
-
-            if (pathsToQuery.length === 1) return { path, value: { $in } };
-
-            const nextSubPath = pathsToQuery[i + 1].path;
-
-            return {
-              value: { [nextSubPath]: { $in } },
-            };
-          }
-
-          const subQuery = priorQueryResult.value;
-          const result = await SubModel.find(subQuery, subQueryOptions);
-
-          const $in = result.map((doc) => doc._id.toString());
-
-          // If it is the last recursion
-          // then pass through the search param
-          if (i + 1 === pathsToQuery.length) {
-            return { path, value: { $in } };
-          }
-
-          return {
-            value: {
-              _id: { $in },
-            },
-          };
-        }, Promise.resolve(initialRelationshipQuery));
-
-        return relationshipQuery;
-      }
-
-      if (operator && validOperators.includes(operator)) {
-        const operatorKey = operatorMap[operator];
-
-        // Some operators like 'near' need to define a full query
-        // so if there is no operator key, just return the value
-        if (!operatorKey) {
-          return {
-            path,
-            value: formattedValue,
-          };
-        }
-
-        return {
-          path,
-          value: { [operatorKey]: formattedValue },
-        };
-      }
-    }
-    return undefined;
   }
 }
